Convert Unix paths in args/env to Windows form for native Win32 apps
Many of our build processes are made up of a mix of Cygwin tools
(makepkg/bash for starters) and native Windows tools. When building
things the paths of input and output files and directories are often
communicated between them via process arguments or environment
variables. The problem here is that those are in many cases not
compatible.

This introduces some "magic" where passing Unix paths to Win32
applications (via command-line arguments and/or environment variables)
are auto-converted to their Win32 variants.

Sometimes, this behavior is undesirable (e.g. when passing regular
expressions via the form that starts and ends in slashes, which this new
logic can mistake for Unix paths). For these instances, there are two
ways to disable the automatic path conversion:

- by setting `MSYS_NO_PATHCONV` to a non-empty string,

- surgically, by setting the `MSYS2_ENV_CONV_EXCL` and
  `MSYS2_ARG_CONV_EXCL` environment variables.

Find a more verbose description at
https://www.msys2.org/docs/filesystem-paths/.

Co-authored-by: Christoph Reiter <reiter.christoph@gmail.com>
Co-authored-by: 마누엘 <nalla@hamal.uberspace.de>
Co-authored-by: Johannes Schindelin <johannes.schindelin@gmx.de>