fixup! Add MSYS2 triplet
We now use "x86_64-pc-cygwin" in MSYS2, so this commit is no longer
doing anything, and removing it would simplify rebasing if upstream
rewrites those files. There is the one hunk which changes the dll prefix
from "cyg" to "msys-", but the runtime doesn't have any DLLs which this
affects currently.

This fixes https://github.com/msys2/msys2-runtime/issues/302.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>